pyproject.toml
[project]
name = "tiny-lm"
version = "0.1.0"
description = "Training small language models from scratch with multiple architectures"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "ferjorosa"}
]
dependencies = [
    "chainlit>=2.9.6",
    "datasets>=4.5.0",
    "gguf>=0.17.1",
    "protobuf>=6.33.5",
    "python-dotenv>=1.2.1",
    "pytorch-lightning>=2.6.1",
    "pyyaml>=6.0.3",
    "rustbpe>=0.1.0",
    "sentencepiece>=0.2.1",
    "tiktoken>=0.12.0",
    "tokenizers>=0.22.2",
    "torch>=2.10.0",
    "trackio>=0.15.0",
    "transformers>=5.0.0",
]

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["tiny_lm*"]

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[tool.uv.sources]
torch = [{ index = "pytorch-cu128" }]
torchvision = [{ index = "pytorch-cu128" }]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "pytest>=8.0.0",
    "ruff>=0.14.14",
    "ty>=0.0.14",
]

[tool.ruff]
line-length = 88
